# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global indirect tax engine API (VAT for Europe, Sales/Use Tax for the Americas)
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryUpdateRateType } from './categoryUpdateRateType';

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  code?: string;
  rateType?: CategoryUpdateRateType;
  /** @nullable */
  description?: string | null;
}