# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global indirect tax engine API (VAT for Europe, Sales/Use Tax for the Americas)
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalNet: number;
  totalTax: number;
  totalGross: number;
  transactionCount: number;
  jurisdictionCount: number;
  reverseChargeCount: number;
  euTax: number;
  usTax: number;
}