# lib/api-zod/src/generated/types/jurisdictionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global indirect tax engine API (VAT for Europe, Sales/Use Tax for the Americas)
 * OpenAPI spec version: 0.1.0
 */
import type { JurisdictionInputRegion } from './jurisdictionInputRegion';
import type { JurisdictionInputTaxType } from './jurisdictionInputTaxType';

export interface JurisdictionInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
  region: JurisdictionInputRegion;
  taxType: JurisdictionInputTaxType;
  standardRate: number;
  /** @nullable */
  reducedRate?: number | null;
  /** @minLength 1 */
  currency: string;
  active?: boolean;
}