# lib/api-zod/src/generated/types/jurisdictionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global indirect tax engine API (VAT for Europe, Sales/Use Tax for the Americas)
 * OpenAPI spec version: 0.1.0
 */
import type { JurisdictionUpdateRegion } from './jurisdictionUpdateRegion';
import type { JurisdictionUpdateTaxType } from './jurisdictionUpdateTaxType';

export interface JurisdictionUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  code?: string;
  region?: JurisdictionUpdateRegion;
  taxType?: JurisdictionUpdateTaxType;
  standardRate?: number;
  /** @nullable */
  reducedRate?: number | null;
  /** @minLength 1 */
  currency?: string;
  active?: boolean;
}